#### Summary
Pileup alignments per chunk to improve the performance

#### Benchmark
##### Pileup many alignments in a short region
```clojure
(time (with-open [r (sam/reader "/path/to/bam/SRP006438.chr1.bam")]
  (dorun (pileup/pileup r {:chr "chr1", :start 152210000, :end 152214999}))))
```

- master: 8523 ms
- PR: 6013 ms (-29.4%)

##### Pileup-bench
- master: `time: 56.355322 sec, sd: 4.858070 ms`
- PR: `time: 55.349883 sec, sd: 7.576505 ms` (-1.8%)

#### Changes
- Read and make chunks which have all pairs
- Correct `quals-at-ref` of an overlapping pair by looking up `quals` in the corrected quals map
- Pileup alignments in each chunk
- Add pileup tests

#### Tests
- `lein check` 🆗
- `lein test :all` 🆗
- `lein eastwood` 🆗