@yito88 2019-04-17
The cause of performance degradation for the longer region was duplicating chunks.
When concatenating chunks for overlapping pairs, it repeatedly piled up for multiple regions including these pairs.

I've tried benchmarking after fixing that.

sample: WES `SRP006438.sorted.bam`

- gene: NRAS `chr1:114703000-114723000`

  |branch|time|
  |:-:|:-:|
  |`master`|`time: 2.424547 sec, sd: 14.857820 µs`|
  |`feature/pileup-chunks`|`time: 2.217902 sec, sd: 129.108257 µs`|

- gene: EGFR `chr7:55015000-55215000`

  |branch|time|
  |:-:|:-:|
  |`master`|`time: 5.902383 sec, sd: 30.402538 µs`|
  |`feature/pileup-chunks`|`time: 5.497433 sec, sd: 109.532839 µs`|